changing structure of handlers
