+++ backend/routes/bot/routes/user/group/left.py
@@ -1,0 +1,26 @@
+from aiogram import Router, F
+from aiogram.types import ChatMemberUpdated
+
+from redis.asyncio import Redis
+from celery.result import AsyncResult
+
+from backend.celery_app.celery_config import celery_app
+
+router = Router()
+
+
+@router.chat_member(
+    (F.old_chat_member.is_member == True) &
+    (F.new_chat_member.is_member == False)
+)
+async def on_user_left(event: ChatMemberUpdated,
+                       redis: Redis) -> None:
+    user_id = event.old_chat_member.user.id
+    chat_id = event.chat.id
+    keys = [key async for key in redis.scan_iter(f"celery:kick:{user_id}:{chat_id}:*")]
+    for key in keys:
+        _, _, _, _, msg_id = key.split(":")
+        result = AsyncResult(key, app=celery_app)
+        result.revoke(terminate=True)
+        await redis.delete(key)
+        await event.bot.delete_message(chat_id=chat_id, message_id=msg_id)
