+++ backend/routes/bot/routes/user/private/callback/confirmation.py
@@ -1,4 +1,5 @@
 from typing import Callable
+
 from aiogram import Router
 from aiogram.types import CallbackQuery
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -7,32 +8,26 @@ from celery.result import AsyncResult
 
 from backend.routes.bot.utils import yes_permissions
 from backend.celery_app.celery_config import celery_app
-from backend.routes.bot.keyboards.callback_factory import ConfirmTelegramUser, Languages
+from backend.routes.bot.keyboards.callback_factory import ConfirmTelegramUser
 from backend.routes.bot.cruds import set_telegram_id, get_telegram_id
 
-router = Router(name="Private callback router")
+
+router = Router()
 
 
 @router.callback_query(ConfirmTelegramUser.filter())
-async def confirmation_buttons(c: CallbackQuery,
-                               callback_data: ConfirmTelegramUser,
-                               db_session: AsyncSession,
-                               redis: Redis,
-                               _: Callable[[str], str]) -> None:
-    """
-        Handles the callback query for Telegram user confirmation.
-        Checks if the user is already authorized by their sub (unique identifier).
-        If the confirmation number matches, updates the user's Telegram ID in the database.
-        If the confirmation number is incorrect, notifies the user.
-        Deletes the confirmation message after processing.
-    """
+async def confirmation_buttons(
+    c: CallbackQuery,
+    callback_data: ConfirmTelegramUser,
+    db_session: AsyncSession,
+    redis: Redis,
+    _: Callable[[str], str]
+) -> None:
 
     if await get_telegram_id(session=db_session, sub=callback_data.sub) is not None:
         await c.message.answer(_("Уже авторизирован!"))
-        await c.message.delete()
-        return
 
-    if callback_data.number == callback_data.confirmation_number:
+    elif callback_data.number == callback_data.confirmation_number:
         user_id = await set_telegram_id(session=db_session,
                                         sub=callback_data.sub,
                                         user_id=c.from_user.id)
@@ -48,15 +43,3 @@ async def confirmation_buttons(c: CallbackQuery,
     else:
         await c.message.answer(_("Введенный вами символ неверный!"))
     await c.message.delete()
-
-
-@router.callback_query(Languages.filter())
-async def choose_language(c: CallbackQuery,
-                          callback_data: Languages,
-                          redis: Redis,
-                          _: Callable[[str], str]):
-    key = f"language:{c.from_user.id}"
-    await redis.set(key, callback_data.language)
-    await c.message.answer(_("Язык успешно изменен!"))
-    await c.message.delete()
-
